+++ lib/common/Utils.cpp
@@ -51,7 +51,7 @@ void SplitString(const std::string &String, char SplitOn, std::vector<std::strin
 	while(e = String.find_first_of(SplitOn, b), e != String.npos)
 	{
 		// Get this string
-		unsigned int len = e - b;
+		size_t len = e - b;
 		if(len >= 1)
 		{
 			rOutput.push_back(String.substr(b, len));
